WIP: A Night at the Opera! :-)
I have imported a FEN-like representation of the famous 'Opera' game, and have
written a utility function to parse it and build the corresponding BlindBoards.
It seems to work! Now I want to diff these boards into moves, throws these moves
to gnuchess, and see if I can replay the whole game through gnuchess, export it,
and make sure that the exported PGN is identical... :-)